[refactor] : 모바일 UI 컴포넌트 폴더 구조 변경 및 적용
