+++ src/stories/mobile/MobileCreateButton.stories.tsx
@@ -1,6 +1,7 @@
+/* eslint-disable no-console */
 import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
-import MobileCreateButton from '@/components/mobile/MobileCreateButton/MobileCreateButton';
+import MobileCreateButton from '@/components/mobile/atom/MobileCreateButton/MobileCreateButton';
 import { BrowserRouter } from 'react-router-dom';
 
 const meta: Meta<typeof MobileCreateButton> = {
