+++ src/stories/mobile/MobileSideMenu.stories.tsx
@@ -2,10 +2,10 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
-import MobileSideMenu from '@/components/mobile/MobileSideMenu/MobileSideMenu';
+import MobileSideMenu from '@/components/mobile/atom/MobileSideMenu/MobileSideMenu';
 import { BrowserRouter } from 'react-router-dom';
 
-const meta = {
+const meta: Meta<typeof MobileSideMenu> = {
   title: 'mobile/MobileSideMenu',
   component: MobileSideMenu,
   parameters: {
@@ -16,6 +16,8 @@ const meta = {
     isOpen: true,
     accessToken: 'true',
     handleLoginButton: () => console.log('로그인 버튼 클릭'),
+    setIsOpen: (isOpen: boolean) =>
+      console.log(`setIsOpen called with: ${isOpen}`),
   },
   decorators: [
     (Story) => (
@@ -24,9 +26,10 @@ const meta = {
       </BrowserRouter>
     ),
   ],
-} satisfies Meta<typeof MobileSideMenu>;
+};
 
 export default meta;
+
 type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {};
